style: rename crawler to collector
