+++ src/main.go
@@ -6,7 +6,7 @@ import (
 	"io/ioutil"
 	"path/filepath"
 
-	"github.com/arasHi87/ScoreboardCrawler/src/crawler"
+	"github.com/arasHi87/ScoreboardCrawler/src/collector"
 	"github.com/arasHi87/ScoreboardCrawler/src/util"
 	"github.com/gocolly/colly"
 	"github.com/sirupsen/logrus"
@@ -16,10 +16,10 @@ var log = logrus.New()
 
 func main() {
 	var users map[string]map[string]string
-	crawlers := map[string]*colly.Collector{
-		"toj":  crawler.TojCrawler(),
-		"uva":  crawler.UvaCrawler(),
-		"tioj": crawler.TiojCrawler(),
+	collectors := map[string]*colly.Collector{
+		"toj":  collector.TojCollector(),
+		"uva":  collector.UvaCollector(),
+		"tioj": collector.TiojCollector(),
 	}
 
 	// load users
@@ -49,22 +49,22 @@ func main() {
 				switch judgeNmae {
 				case "toj":
 					url := fmt.Sprintf("https://toj.tfcis.org/oj/be/chal?off=0&proid=%s&acctid=%s", problemId, userId)
-					crawlers[judgeNmae].Request("GET", url, nil, ctx, nil)
+					collectors[judgeNmae].Request("GET", url, nil, ctx, nil)
 				case "uva":
 					url := fmt.Sprintf("https://uhunt.onlinejudge.org/api/p/num/%s", problemId)
 					ctx.Put("pnum", problemId)
-					crawlers[judgeNmae].Request("GET", url, nil, ctx, nil)
+					collectors[judgeNmae].Request("GET", url, nil, ctx, nil)
 				case "tioj":
 					url := fmt.Sprintf("https://tioj.ck.tp.edu.tw/submissions.json?filter_username=%s&filter_problem=%s", userId, problemId)
-					crawlers[judgeNmae].Request("GET", url, nil, ctx, nil)
+					collectors[judgeNmae].Request("GET", url, nil, ctx, nil)
 				}
 			}
 		}
 	}
 
 	// get all submission
-	for _, crawler := range crawlers {
-		crawler.Wait()
+	for _, collector := range collectors {
+		collector.Wait()
 	}
 
 	// integration all result into result.json
